backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
gradio_client==1.3.0
gunicorn==22.0.0
